requirements.txt
pandas==2.2.2
matplotlib==3.9.2
seaborn==0.13.2
numpy==1.26.4
requests==2.31.0
python-dotenv==1.0.1
scikit-learn==1.5.2
xgboost==3.0.4
optuna==4.5.0
joblib==1.4.2
wordcloud==1.9.4
